refactor(#94): modify repository
