+++ src/hot-posts/repositories/hot-posts.repository.ts
